- Split main menu in main & more menus, - Simplify options, - Decimates old positions when not sent to avoid crashes (overflow), - Minor changes.

+++ vgpsmobile/src/fr/victorb/mobile/vgps/gps/BluetoothGps.java
@@ -127,8 +127,8 @@ public class BluetoothGps extends Gps implements Runnable {
                         position.time.second = (byte)Integer.parseInt(string.substring(4, 6));                    
                         position.latitude = Float.parseFloat(split.next()) * (split.next().toUpperCase().equals("N")?1:-1) / 100;
                         position.longitude = Float.parseFloat(split.next()) * (split.next().toUpperCase().equals("E")?1:-1) / 100;
-                        position.elevation = elevation;                    
-                        position.speed = Float.parseFloat(split.next()) * 1.852f;  // speed (knots -> km/h)    
+                        position.elevation = (short)elevation;                    
+                        position.speed = (byte)(Float.parseFloat(split.next()) * 1.852f);  // speed (knots -> km/h)    
                         split.next();                                              // heading 
                         string = split.next();
                         position.date.day = (byte)Integer.parseInt(string.substring(0, 2));
